pyproject.toml
[project]
name = "smart-portfolio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "keras>=3.13.2",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "tensorflow>=2.21.0",
    "yfinance>=1.2.0",
]
